Create bar.c to set bar functions
