+++ appinfo.json
@@ -37,19 +37,7 @@
             },
             {
                 "file": "fonts/square.ttf",
-                "name": "FONT_SQUARE_34",
-                "targetPlatforms": null,
-                "type": "font"
-            },
-            {
-                "file": "fonts/square.ttf",
-                "name": "FONT_SQUARE_14",
-                "targetPlatforms": null,
-                "type": "font"
-            },
-            {
-                "file": "fonts/square.ttf",
-                "name": "FONT_SQUARE_26",
+                "name": "FONT_SQUARE_18",
                 "targetPlatforms": null,
                 "type": "font"
             },
@@ -61,7 +49,19 @@
             },
             {
                 "file": "fonts/square.ttf",
-                "name": "FONT_SQUARE_18",
+                "name": "FONT_SQUARE_26",
+                "targetPlatforms": null,
+                "type": "font"
+            },
+            {
+                "file": "fonts/square.ttf",
+                "name": "FONT_SQUARE_14",
+                "targetPlatforms": null,
+                "type": "font"
+            },
+            {
+                "file": "fonts/square.ttf",
+                "name": "FONT_SQUARE_34",
                 "targetPlatforms": null,
                 "type": "font"
             }
