--- src/battery.h
@@ -1,8 +1,0 @@
-#include <pebble.h>
-#pragma once
-
-Layer *batt_layer, *static_layer;
-
-void battery_handler(BatteryChargeState state);
-void batt_layer_draw(Layer *layer, GContext *ctx);
-void static_layer_draw(Layer *layer, GContext *ctx);
